final ver 3, small changes
