--- .idea/.gitignore
@@ -1,7 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
-.idea
-/out
-
-
