+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/ava111.iml" filepath="$PROJECT_DIR$/ava111.iml" />
+      <module fileurl="file://$PROJECT_DIR$/CRUD-Pantil.iml" filepath="$PROJECT_DIR$/CRUD-Pantil.iml" />
     </modules>
   </component>
 </project>
