--- src/com/javarush/ava111/Service/ApplicationSettings.java
@@ -1,7 +1,0 @@
-package com.javarush.ava111.Service;
-
-public class ApplicationSettings {
-    public static final String FILE_PATH = "C:\\Users\\whale\\OneDrive\\Рабочий стол\\Флешка\\Source.txt";
-    public static final String DATE_PATTERN = "dd/MM/yy";
-    public static final String OUTPUT_FILE_PATH = "C:\\Users\\whale\\OneDrive\\Рабочий стол\\Флешка\\output.txt";
-}
