--- src/com/javarush/ava111/Service/Util.java
@@ -1,9 +1,0 @@
-package com.javarush.ava111.Service;
-
-import java.text.SimpleDateFormat;
-
-import static com.javarush.ava111.Service.ApplicationSettings.DATE_PATTERN;
-
-public class Util {
-    public static final SimpleDateFormat formatter = new SimpleDateFormat(DATE_PATTERN);
-}
