+++ src/main/java/com/osokin/config/ApplicationSettings.java
@@ -1,0 +1,10 @@
+package com.osokin.config;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+public class ApplicationSettings {
+    public static final String FILE_PATH = "C:\\Users\\whale\\IdeaProjects\\ava111\\src\\main\\resources\\source.txt";
+    public static final String DATE_PATTERN = "dd/MM/yy";
+    public static final String OUTPUT_FILE_PATH = "output.txt"; //TODO classpath
+}
