+++ src/main/java/com/osokin/service/DtoConverter.java
@@ -1,21 +1,17 @@
-package com.javarush.ava111.Service;
+package com.osokin.service;
 
-import com.javarush.ava111.Dto.CreateDto;
-import com.javarush.ava111.Dto.DeleteDto;
-import com.javarush.ava111.Dto.ReadDto;
-import com.javarush.ava111.Dto.UpdateDto;
+import com.osokin.dto.CreateDto;
+import com.osokin.dto.DeleteDto;
+import com.osokin.dto.ReadDto;
+import com.osokin.dto.UpdateDto;
 
 import java.text.ParseException;
 import java.util.logging.Logger;
 
-import static com.javarush.ava111.Service.Util.*;
+import static com.osokin.service.Util.*;
 
 public class DtoConverter {
     private static Logger logger = Logger.getLogger(DtoConverter.class.getName());
-
-
-
-
     public static CreateDto createDto(String[] params) throws ParseException {
         CreateDto createDto = new CreateDto();
         createDto.setName(params[1]);
@@ -25,10 +21,8 @@ public class DtoConverter {
         createDto.setPassportIssueDate(formatter.parse(params[5]));
         createDto.setCity(params[6]);
 
-
         return createDto;
     }
-
     public static UpdateDto updateDto(String[] params) throws ParseException {
         UpdateDto updateDto = new UpdateDto();
         updateDto.setID(Integer.parseInt(params[1]));
