+++ src/main/java/com/osokin/service/Util.java
@@ -1,0 +1,9 @@
+package com.osokin.service;
+
+import com.osokin.config.ApplicationSettings;
+
+import java.text.SimpleDateFormat;
+
+public class Util {
+    public static final SimpleDateFormat formatter = new SimpleDateFormat(ApplicationSettings.DATE_PATTERN);
+}
