+++ src/main/resources/source.txt
@@ -1,0 +1,17 @@
+create	Стас	23	5014	121212	12/10/14	Новосибирск
+create	Виктор	13	5015	131313	12/11/20	Казань
+create	Антон	10	5010	100000	12/10/10	Москва
+create	Александр	43	5011	567909	12/08/14	Бердск
+update	1	Стас	23	5016	121212	12/10/14	Улан-Удэ
+update	2	Виктор	13	5012	131313	12/11/08	Нижний Новгород
+update	3	Антон	19	5010	100222	12/10/11	Нижневартовск
+update	4	Александр	43	5011	547959	12/08/15	Астана
+delete	1	3	8
+create	Михаил	20	5015	125099	12/06/16	Махачкала
+create	Виталий	43	5011	567909	12/08/14	Бердск
+create	Алексей	15	5011	567909	12/08/14	Бердск
+create	Сергей	43	5011	567909	12/08/14	Бердск
+create	Родион	18	5011	567909	12/08/14	Бердск
+create	Дмитрий	43	5011	567909	12/08/14	Бердск
+create	Алена	43	5011	567909	12/08/14	Бердск
+create	Владислава	43	5011	567909	12/08/14	Бердск
